Move task navigation keybinding to  plus some additional cleanup
